segregates help command usage
